requirements.txt
waitress==1.4.4
gunicorn==20.0.4

alembic==1.3.3
Flask==1.1.1
Flask-Compress==1.4.0
Flask-Cors==3.0.8
Flask-HTTPAuth==3.3.0
Flask-JSON==0.3.4
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
GeoAlchemy2==0.6.3
marshmallow==3.3.0
marshmallow-sqlalchemy==0.21.0
SQLAlchemy==1.3.13

psycopg2==2.8.4
psycopg2-binary==2.8.5

Werkzeug==1.0.0

aniso8601==8.0.0
attrs==19.3.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
distro==1.4.0
docutils==0.15.2
ecdsa==0.15
gevent==1.4.0
GPUtil==1.4.0
greenlet==0.4.15
idna==2.8
importlib-metadata==1.4.0
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
joblib==0.14.1
Mako==1.1.1
MarkupSafe==1.1.1
more-itertools==8.1.0
opencage==1.2
packaging==20.1
pluggy==0.13.1
py==1.8.1
pyasn1==0.4.8
pyparsing==2.4.6
pyssdb==0.4.2
python-dateutil==2.8.1
python-dotenv==0.10.5
python-editor==1.0.4
python-jose==3.1.0
pytz==2019.3
pyzmq==19.0.0
requests==2.22.0
rsa==4.0
sentry-sdk==1.1.0
six==1.14.0
termcolor==1.1.0
tqdm==4.45.0
urllib3==1.25.8
wcwidth==0.1.8
zipp==2.1.0
tldextract==2.2.3
django-environ==0.4.5


# no version
scikit-learn
scipy
numpy

parsel==1.6.0

google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-bigquery
google-cloud-core
google-cloud-firestore
google-cloud-storage
google-resumable-media
googleapis-common-protos

# pybloom-live
# https://github.com/joseph-fox/python-bloomfilter
pybloom-live

warcio==1.7.4

psutil==5.7.3

XlsxWriter==1.3.7

# AWS stuff
boto3
fastavro
PyAthena
botocore
boto3-stubs
tenacity
s3transfer

# tests
pytest
flake8
pyflakes
